+++ src/components/index.js
@@ -1,0 +1,4 @@
+import Svg from './Svg'
+import Rect from './Rect'
+
+export { Svg, Rect }