+++ src/svgx.js
@@ -1,4 +1,5 @@
 import render from './render'
 import Svg from './components/Svg'
+import Rect from './components/Rect'
 
-export { render, Svg }
+export { render, Svg, Rect }